pyproject.toml
[tool.poetry]
name = "time-tracker"
version = "0.1.0"
description = ""
authors = ["Jose Cabeda <jecabeda@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
tt = "time_tracker.main:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.1.0"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
flake8 = "^3.7.9"
mypy = "^0.770"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
